27.SmallData/requirements.txt
pymongo[srv]
python-dotenv
pymongoarrow
pandas
numpy
pyarrow
deltalake
duckdb
streamlit
plotly.express